ucits-bot/requirements.txt
boto3==1.26.41
botocore==1.29.41
certifi==2022.12.7
charset-normalizer==2.1.1
et-xmlfile==1.1.0
idna==3.4
jmespath==1.0.1
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.13
